backend/requirements.txt
Flask==2.3.2
pdfplumber==0.9.0
pandas==2.0.3
mysql-connector-python==8.0.33
werkzeug
gunicorn